% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaintegerPopulationFactory.R
\name{gaintegerPopulationFactory}
\alias{gaintegerPopulationFactory}
\title{GA Integer Population Factory}
\usage{
gaintegerPopulationFactory(zeroFraction, verbose = FALSE)
}
\arguments{
\item{zeroFraction}{Numeric value between 0 and 1 specifying the fraction of variables
to set to zero in each individual}

\item{verbose}{Logical indicating whether to print information about zero fraction}
}
\value{
A function that takes a GA object and returns an initial population matrix
}
\description{
Creates a factory function for generating initial populations for genetic algorithms
with integer chromosomes, where a specified fraction of variables are set to zero.
}
