% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summaryComponents.R
\name{print.summaryComponents}
\alias{print.summaryComponents}
\title{Print method for summaryComponents}
\usage{
\method{print}{summaryComponents}(x, showDetails = FALSE, maxLinesDetails = 200L, ...)
}
\arguments{
\item{x}{An object produced by \code{summaryComponents()}.}

\item{showDetails}{Logical. If TRUE, also print row/column indices per component.}

\item{maxLinesDetails}{Integer. Max number of indices to print per dimension
(rows/cols) before truncation (default 200).}

\item{...}{Unused.}
}
\description{
Print method for summaryComponents
}
